Add log functionality
log to a file as well as standard out, we're using logger from gpbackup
for now.

Refactor code to add integrational test to start a server.
For timing, use a client to ensure that we've started the server before
testing. Since the server is started on a separate thread, we coudln't
properly stop the server unless we had a client.

Signed-off-by: Marbin Tan <mtan@pivotal.io>