WIP: expand: adopt new cluster size at begin of xact
When new segments are added to the cluster, existing transactions will
still work with the old cluster size so their gangs are only created on
the old segments and motions are also limited on them.  Only new
transactions will get the new cluster size and work with the new
segments.

Co-authored-by: Jialun Du <jdu@pivotal.io>
Co-authored-by: Ning Yu <nyu@pivotal.io>