WIP: expand: add a global lock to protect catalog changes
Catalog changes must be prevented during online expanding.  A global
lock is introduced for this purpose, all updates to catalog (except some
master only tables) will hold this lock in shared mode, the online
expand operations should only be made after holding this lock in
exclusive mode, then all existing catalog changes are commited or
rollbacked.

A special case is as below:

	tx1: begin;
	tx2: online expand;
	tx1: create table t1 (c1 int);

A transaction begins before online expand, but attempt to update catalog
after online expand, how could we handle that?  This transaction will
still see the old cluster size and dispatch the create table command to
old segments, then t1 is not created on the new segments.

To prevent this inconsistency at present we have to raise an error for
catalog changes in existing transactions after the online expand.

Co-authored-by: Jialun Du <jdu@pivotal.io>
Co-authored-by: Ning Yu <nyu@pivotal.io>