Fix INSERT ... VALUES.
When cleaning up the source code for PR we disabled an adjustment on
numsegments, which causes INSERT ... VALUES to generate below plan:

    explain insert into t1 (c1, c2) values (1, 0);
                               QUERY PLAN
    ----------------------------------------------------------------
     Insert on t1
       ->  Redistribute Motion 3:1  (slice1; segments: 3)
             Hash Key: c1, c2
             ->  Result
     Optimizer: legacy query optimizer
    (5 rows)

A redistribute motion was added, INSERT will thus hang there.

Add back the adjustment in adjust_modifytable_flow(), put General node
on the same segments with target relation to prevent the motion being
added.  So we generate correct plan again:

    explain insert into t1 (c1) values (1);
                       QUERY PLAN
    ------------------------------------------------
     Insert on t1
       ->  Result
     Optimizer: legacy query optimizer
    (3 rows)