explain: show correct Segments when format is yaml/json.
When explain format is text we only display Segments for mation nodes,
and its value is fetched with the assumption that the plan is Motion.

However Segments is displayed for all nodes when explain format is
yaml/json, then incorrect values are displayed.

Updated the logic to fetch correct Segments for all nodes.