Make Null locus with proper numsegments.
Null locus are usually made as a placeholder, which will later be remade
to an actual non-null locus, in such a case we could assign any value as
numsegments to Null locus.  We used to set 0 to it, now use EVIL
instead.

There is also cases that numsegments of Null locus is not meaningless,
these cases are not touched.

Also updated some places that set numsegments to getgpsegmentCount()
directly, use ALL instead.