Gather to ALL for Append with SegmentGeneral.
Suppose t1 is a hash distributed table on 1 segment, d2 is a replicated
table on 2 segments, consider below query:

    select * from t1 join all select * from d2;

Append node with at least one SegmentGeneral subpath will gather all the
subpaths to one segment, it's important to ensure that all the subpaths
can be gathered to the SAME segment, we must set the same numsegments
for all the SingleQE, there are many options:

1. a constant 1;
2. Min(numsegments of all subpaths);
3. Max(numsegments of all subpaths);
4. ALL;

Options 2 & 3 need to decide the value with an extra scan, option 1 puts
all the SingleQE on segment 0 which makes segment 0 a bottle neck.  So
we choose option 4, ALL helps to balance the load on all the segments
and no extra scan is needed.