Fix update the same row concurrently bug.
Splitupdate Node was introduced to support update hash cols
of hash-distributed tables. This technique depends that
we acquire ExclusiveLock on the tables. With the commit of
GDD, we do not upgrade lock for heaptable's update/delete
operation, which breaks the assumption of splitupdate, which
will leads to more tuples when concurrently update the
same row's hash columns.

We fix this by lock the table in ExclusiveMode in ExecInitSplitUpdate
on QD to make thing correct but lose the performance of concurrency.

We should refactor this part of code to improve concurrency performance.

Co-authored-by: Zhenghua Lyu zlv@pivotal.io
Co-authored-by: Shujie Zhang shzhang@pivotal.io