Adjust gang size according by numsegments
Now we have  partial tables and flexible GANG API, so we can allocate
GANG according to `numsegments`.

With the commit 4eb65a53, gpdb supports table distributed on partial segments,
and with the series of commits (a3ddac06, 576690f2), gpdb supports flexible
gang API. Now it is a good time to combine both the new features. The goal is
that creating gang only on the necessary segments for each slice. This commit
also improves singleQE gang scheduling and does some code clean work. However,
if ORCA is enabled, the behavior is just like before.

The outline of this commit is:
  * Modify the FillSliceGangInfo API so that gang_size is truly flexible.
  * Remove numOutputSegs and outputSegIdx fields in motion node. Add a new
    field isBroadcast to mark if the motion is a broadcast motion.
  * Remove the global variable gp_singleton_segindex and make singleQE
    segment_id randomly(by gp_sess_id).
  * Remove the field numGangMembersToBeActive in Slice because it is now
    exactly slice->gangsize.
  * Modify the message printed if the GUC Test_print_direct_dispatch_info
    is set.
  * Explicity BEGIN create a full gang now.