Remove the distributed log on new segments
df19119c208 eliminate distributed transaction log creation
and maintenance on QD (Only a 32K `pg_distributedlog/0000`
 file exists). Gpexpand will copy data files from
QD to new segments, So in the new segments the oldestXID
 is 3 (loaded from pg_controlfile which copied from QD).

After new segments join the cluster, they will maintain the
oldestXmin, so it will loop to find the page which is in the
distributed transaction log.

If the local transaction ID (xid) is huge on new segments, it will
lead to a hole between 0000 and TransactionIdToPage(xid),
then an error will be raised.

In this commit we truncate the distributedlog with cutoff of oldestxid
on new segments,  then the hole is gone, so the oldestXmin will be
initialized to oldestLocalXmin.